Handles errors when loading batch tasks from multiple ids
I ran into an error when I used `operator.add` as a batch_method. This
fails because operator.add has the wrong signature. The problem is that
this failure kills the worker rather than just failing that particular
job. In order to allow the rest of the pipeline to continue, we trap the
error just like when we fail loading a single task.