Sends an error message to the scheduler upon failing to load tasks
Since moving to an assistant-based workflow, assistants often pick up
tasks that users are developing in their own branches. As assistants have
no way of running these, we get a fair amount of failures. Unfortunately,
the scheduler never receives the reason for these failures, leaving many
people confused as to what exactly happened. Luckily, there's an easy fix
for this by simply adding the traceback when informing the scheduler of
the failure.