Enables running with 0 workers to support assistants (#1861)
In order to allow assistants to take care of running all tasks, we need a
way to run scheduling with 0 workers. This adds a slightly tweaked worker
code path to allow a process with 0 workers to stay alive while skipping
any get_work calls to the scheduler. Instead, the worker makes a
count_pending call to see if it's still necessary.