Prevents batched tasks from all getting marked RUNNING
I've recently seen some of my workers simultaneously running multiple
tasks from the same batchable class, when the should all be running
together as a single batch. Furthermore, these tasks exceeded both
resource constraints and the number of workers supposed to be running at
the same time.

After checking the logs, I saw that the worker would actually run these
one at a time upon receiving each one from the scheduler. My best guess
after looking into it for a week is that the worker does not receive a
response from the scheduler after marking the batch runner as RUNNING,
and the scheduler updates the entire batch's status on the retry. I
verified that this is indeed a possibility with a unit test and have
patched that bug here.