Allow a file to be moved cross-device.
With `os.rename` you get a `OSError: [Errno 18] Invalid cross-device link`
if you attempt to move a file across filesystems.

`shutil.move` covers the basic case:

> If the destination is on the current filesystem, then os.rename() is used.

and additionally handles the cross-device case.

If the destination is on a different filesystem,
we approximate atomicity by first copying the source
to a temporary file on the target filesystem, followed
by a rename. Finally the source file is removed.