Prevent stowaway resources in scheduler
When batching tasks together, we implicitly assume that all tasks being
batched have the same resource limitation and that the batched task will
have the same resource. In order to make this assumption more explicit,
we now enforce that all tasks returned in a batch group have the same
resources.