Cleanup code and adhere to E305 (#1922)
This patch started out as fixing the red in the current travis flake8
step.  As most occurrences was about simply removing the over-(mis)-used
`if __name__ == "__main__"` pattern that was way to prevalent in the luigi
code base.

This is kind of analogous to the other time flake8 suddenly got more
sensitive. See spotify/luigi#1786