Only run ecs_test.py if nose attribute awsecs is set
The `test/contrib/ecs_test.py` requires AWS credentials and an AWS account
with a running ECS cluster to execute properly. Previously, it's been
prevented from running by being skipped unless `boto3` has been manually
installed in the test environment.

To be able to include `boto3` in the default test environment, and still
prevent `ecs_test.py` from running e.g. in CI, this change sets an `awsecs`
[nose attribute][1] on the test, so it will only execute e.g. if
`NOSE_ATTR=awsecs` is in the shell environment, or if tox is invoked
with the `-awsecs` tox-environment, e.g. like

    tox -e py27-awsecs

Specifically, it should not be executed by the Travis CI engine (even if
`boto3` is installed).

[1]: https://nose.readthedocs.io/en/latest/plugins/attrib.html