Allow namespaces to be set to python module
This have been a problem I've thought about for years. And also
consciously gradually refactored the code base in previous patches
towards making this relatively short path.

Coming up with the final syntax and implementation also took me a while.
Initially I imagined there would be a [config] parameter one could set,
or specify by command line. But it turned out to be quite troublesome:

 * It's not clear how much it should affect. Should it also affect the
   things luigi exports?
 * It's hard to transition into it. Imagining a company with independent
   teams, the `auto_namespace` function can be applied package by
   package, luigi-version by luigi-version to whoever chooses to apply
   this new convention.

This solution came out to be pretty clean, and with readable tests and
detailed and a bit opinionated documentation.