Set authors to "The Luigi Authors"
As of there being close to 300 commiters to luigi, and today's main
contributors to the code base are not the two original founders anymore
(which is a good sign of success :)). It makes more sense to mention
a generic term as the authors of the software.

I took the name from the Tornado docs (obviously replacing "Tornado"
with "Luigi"). The original authors are still mentioned in the bottom of
the README.md.