Add method rename_dont_move to LocalFileSystem (#2057)
When the method temporary_path is called on LocalTarget, a warning is
given because the default implementation of the method rename_dont_move
is used. However LocalFileSystem has already a more sophisticated
implementation of the desired functionality in the `move` method. Here
rename_dont_move is implemented by means of this `move` method.

Additionally the exception raised by `move` is changed from RuntimeError
to FileAlreadyExists, as the latter is expected by the test of
rename_dont_move.