travis: Run less builds
The best solution would be to actually separate out the hdfs and other contrib parts of luigi to seperate repos. But that ain't going to happen with current work power. But we need fast builds and fewer intermittent errors, we remove a couple of builds from the matrix.

This is also a way to be grateful and respectful for the free Travis build resources.