Defer S3 Client instantiation until needed
Previously the S3 Client was instantiated in the `__init__` method of
S3Target, which meant that if you had to, say, traverse a large task
graph, quickly creating and throwing away targets, you'd end up
incurring a lot of wasted connections and slowness.

Now we create the client only when we want to make a remote request.

This does not change any of the semantics of how the target is
instantiated (still can override `s3` property and the `Key` property),
except that making an STS connection is also deferred until the moment
it's actually used.