Check whether renames worked in atomic hdfs pipes
Using the snakebite client, renaming a file to an existing file fails
without raising an exception. Instead, it returns a list of moves and
whether each ones succeeded. This means that when using the snakebite
client, atomic writes to existing files fail silently versus failing
while throwing an error with the hadoopcli client.

There are two ways to fix this. First, we can simply remove the target
file if it already exists. Second, we can ensure that we read the
snakebite error and throw it if it happens. This PR does both.

For atomic directory writes, this also adds a check for whether the
directory was renamed to the target or moved inside an existing target
directory.