Preserve options in svg graph links (#2132)
When clicking on the nodes in an svg graph, a graph will load with that
node as the root but with hide done and show upstream dependencies
unset. This can be very annoying when you suddenly have to load a much
larger graph or your graph flips orientation.

In order to preserve these and any future options, we pass the hash with
taskId stripped out of it into the dependency graph so it can use that
as a basis for the node links. This preserves the options present when
the graph was created, which should match the current options on the
page.

D3 graphs also suffer from the same bug, but they are not addressed by
this change.