Preserve graph options when clicking d3 nodes (#2139)
When clicking a d3 node, the new url was built up with a static set of
options, clearing out things like hide done and invert. This means the
graph would flip or gain lots of nodes when you click a link. In order
to better preserve the options, we now simply change the taskId in the
current URL.