Prevent worker from running same task repeatedly (#2207)
If the scheduler keeps giving the worker the same task_id from get_work,
the worker will just overwrite the same entry in _running_tasks and
never think that it is running too many processes.

I've had to reboot one of my worker machines almost daily due to
hundreds of worker processes being spawned to run the same task. This
seems to be mostly happening with batch tasks, probably because they
require more rounds of communication before the scheduler understands
that the task is running. Since this tends to happen only in short
bursts, I've also added a sleep to give the scheduler a chance to
recover.