Set utf-8 encoding in the MIMEText constructor (#2210)
* Set utf-8 encoding in the MIMEText constructor

If this is not set during the constructor, an exception is thrown in
Python 2 when the e-mail body contains unicode.

* Parse e-mail in WorkerEmailTest

With base64 encoding, the e-mail tests fail trying to validate subject
and body contents. Using email.parser allows us to safely decode
whatever encoding is used and verify the subject and body separately,
resulting in slightly better validation as well.