Speed up task_list when beyond limit (#2239)
When the number of tasks get into the millions, even refreshing the
visualizer can take a minute or more, causing havoc in the pipeline.
Since all we really want in these situations is the counts, we can skip
the more expensive bits of computation and just return the sizes. This
prevents doing upstream checks, but saves a lot of time.

We may want to institute a higher threshold so we can get upstream
numbers if you're only a little above the limit for returning all
tasks.