Hide re-enable and forgive buttons on success
After clicking re-enable, the whole page refreshes. After clicking
forgive failures, nothing happens. Both of these responses are a bit
extreme. To find a happier medium, this updates both buttons to simply
disappear after the response is received from the server. This makes it
easier to do multiple disables and allows for more confidence in
forgiving failures.

This also adds the ability to hold alt with the re-enable to get an
update like with forgive failures.