Optional parameter (#2295)
* Add an OptionalParameter parameter class

This allows a parameter to be a string or None. None is serialized as
an empty string, and empty strings parse as None. This is useful for
luigi.Config classes where we want a None value for some parameters
and don't want warnings about the type being wrong.

* Use OptionalParameter when default is None

This will reduce warnings in the scheduler. It should be safe unless
there are cases where the empty string needs to be distinguished from
None. That doesn't seem to be the case anywhere here.