Only run code coverage cheks for core luigi
I often find that in PRs, codecoverage says that coverage has degraded. Many of the times it's unreliable and also I think we don't need to be so strict for the whole code base. With this change, we'll only check for core (which I hope have no intermittent-ency). 

The number 92.5% is the current coverage according to my calculation. I think it's good to not go below that.