Add force_multiprocessing config to worker plus test and docs. (#2401)
At the moment, luigi uses multiprocessing for tasks only when the number of requested workers is > 1. In some situations, it might be helpful if this was also the case for exactly 1 worker (e.g. when the executed code does something nasty with the main process, TensorFlow in our case). I added a worker config force_multiprocessing which, when True, forces multiprocessing to be used independent of the number of workers. I also added a few lines to the docs.