Add table constraints support to Redshift
While the table keys constraints (UNIQUE, PRIMARY, FOREIGN) aren't
enforced by Redshift, it's still useful to be able to set that data.
Given the situation that there is a monitoring task that manually checks
if the table constraints are respected, it's useful to be able to set
multiple primary keys on a given table.

Based on the syntax of the CREATE commands of Redshift (see link down
below) we think it's useful to add support to it.

https://docs.aws.amazon.com/redshift/latest/dg/r_CREATE_TABLE_NEW.html