Add schema to DB if not exists for Redshift
This adds the functionality to allow for the schema to be created if we
find a schema as part of the table name.

Based on the current implementation, code exists to create the table if
it doesn't exists. It's a fair assumption that we would like to have the
same behavior for schemas.

Without this feature, the task could potential crash and throw an error
saying that it cannot create the table under the related schema if a
schema is specified in the table name and that it doesn't exists.