Clarify "dependencies" in summary
The phrase "present dependencies" tends to confuse users because they think of the "root" task as "not a dependency". So in an output like this:

===== Luigi Execution Summary =====

Scheduled 1 tasks of which:
* 1 present dependencies were encountered:
   - 1 CaseEntity(when=2018-05-14)

Did not run any tasks
This progress looks :) because there were no failed tasks or missing external dependencies

===== Luigi Execution Summary =====

they are misled into thinking that their task's dependency was present, and that Luigi erroneously didn't run the task itself.