Add `target_table` to `PostgresTarget.exists()`
Adding `target_table` to the `exists` method of PostgresTarget checks if a specific `update_id` exists for a specific table, rather than any table.
This way, one can use a "general" update_id (e.g. update_id = '2018-05-01', if a Pipeline has to be executed on a daily basis) for every task involved in this Pipeline.
Since a PostgresTarget can only point to one specific table anyway, it makes sense to constraint its `exists` method to check for an `update_id` of this specific table.

In the current implementation, every task in a Pipeline has to have a specific `update_id`. 
This often results in the table name being integrated to the `update_id` (e.g.: 'table-a-2018-05-01', 'table-b-2018-05-01', ...), which brings redundancy, since the table name is in the column `target_table` anyway.