Add metadata columns to the RDBMS contrib (#2440)
* Add metadata column feature to Redshift

The goal of this feature is to allow metadata column to exists for
specific tables created by the Redshift contrib related tasks.

Given the scenario where we would always have to have a `created_tz`
column at the end of every tables generated by that contrib we could do
the following:

```python
UpdateTableTask(redshift.S3CopyToTable):

    def metadata_columns(self):
        return [('created_tz', 'TIMESTAMP')]

    def post_queries(self):
        return query = 'UPDATE {0} ' \
                       'SET created_tz = CURRENT_TIMESTAMP ' \
                       'WHERE created_tz IS NULL'.format(self.table)
```

Adding layer of abstraction over this feature, you could easily
add many default behavior for specific tables for versioning the table
and more.

This feature is opt-in by default since we don't want this break other
people's pipeline after integrating this.

* Move the Metadata Columns implementation to the RDBMS

As suggested in the code-review, there are multiple other DBs that could
benefit from this change. Currently, only PSQL and Redshift implements
RDMS but other may implement this class and inherits that new behavior.

* Add tests for the Metadata behaviors

We've been internally using this feature for Redshift but moving this to
the RDBMS contrib and adding this behavior to PSQL could have unexpected
side effects, this takes care of testing if the feature works correctly
under Redshift and PSQL.

* Add additional documentation on how to use the new mixin

* Raise ValueError on invalid metadata_columns for RDBMS

If the count of metadata_columns is 0 and we're expecting to add them to
the table, then we raise an error because that is an invalid flow.

The contributor is required to have metadata_columns values if we want
to add that column to the table.