Remove connection from touch call
Since Snowflake heavily use these statements:

* USE ROLE
* USE DATABASE
* USE WAREHOUSE

It's possible that a task sets these value to something else for a given
connection and that we try to write to the marker table with a role that
doesn't have access to it.

If we don't pass the connection object to the touch function, then it
will re-create the connection with the sane defaults.