All configuration parameters in docs now use underscore in their names for consistency.
Previosly it was hard to search some parameters on page because in some places a minus sign
was used and in others is underscore.

This commit fixes this issue.

Also, it fixes few places where one backquote was used to denote a parameter name or value.
RestructuredText format uses two backquotes to denote an inline code block:

https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#inline-markup