remove StreamingBodyAdaptor that didn't stream by chunks
This StreamingBodyAdaptor caused performance issues:
Calling `__iter__()` on a StreamingBodyAdaptor object defaulted to chunks of 1 byte, which was really slowing
down reads. I wasn't able to pass the `size` parameter to change the chunk size

botocore StreamingBody implementation performs better:
https://github.com/boto/botocore/blob/master/botocore/response.py#L90

It defaults to chunks of 1024 bytes, and can be adjusted.
It also offers iter_lines() and iter_chunks() methods