Remove Asana from Luigi users (#2945)
Asana is no longer using Luigi, so for the sake of accuracy, it was suggested that we should be removed from the list of users.  The blog post that describes our use of Luigi did not include any implementation details or tips that would be useful to the broader community of Luigi users.